requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via nltk
cycler==0.11.0
    # via matplotlib
filelock==3.7.1
    # via
    #   huggingface-hub
    #   transformers
fonttools==4.33.3
    # via matplotlib
huggingface-hub==0.8.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.3
    # via requests
joblib==1.1.0
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.4.3
    # via matplotlib
matplotlib==3.5.2
    # via -r requirements.in
nltk==3.7
    # via sentence-transformers
numpy==1.23.0
    # via
    #   matplotlib
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
packaging==21.3
    # via
    #   huggingface-hub
    #   matplotlib
    #   transformers
pillow==9.1.1
    # via
    #   matplotlib
    #   torchvision
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.6.2
    # via
    #   nltk
    #   transformers
requests==2.28.0
    # via
    #   huggingface-hub
    #   torchvision
    #   transformers
scikit-learn==1.1.1
    # via sentence-transformers
scipy==1.8.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via -r requirements.in
sentencepiece==0.1.96
    # via sentence-transformers
six==1.16.0
    # via python-dateutil
threadpoolctl==3.1.0
    # via scikit-learn
tokenizers==0.12.1
    # via transformers
torch==1.11.0
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.12.0
    # via sentence-transformers
tqdm==4.64.0
    # via
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   transformers
transformers==4.20.1
    # via sentence-transformers
typing-extensions==4.2.0
    # via
    #   huggingface-hub
    #   torch
    #   torchvision
urllib3==1.26.9
    # via requests
